+++ CostIncomeCalculator.api/Dtos/IncomeDtos/MonthIncomeDto.cs
@@ -3,6 +3,6 @@ namespace cost_income_calculator.api.Dtos.IncomeDtos
     public class MonthIncomeDto
     {
         public string Category { get; set; }
-        public double IncomeSum { get; set; }
+        public decimal IncomeSum { get; set; }
     }
 }
